requirements.txt
numpy==1.23.2
pandas==2.2.1
plotnine==0.12.4
plotnine_prism==0.3.0
scipy==1.12.0
torch==2.0.1
pytest==7.4.0
pyyaml==6.0.1
matplotlib>=3.7
